Fixed #41
A new incompatibility presented itself after I updated shiny.
In short, ggplotly() was not recognizing my call to
ggplot2::coord_flip() so the forest plots were displaying incorrectly.
It turns out that this is a known issue. Decided to just go
with the flow and display the plots on the default dimension
of ggplot2::geom_boxplot().